robotic_dog/requirements.txt
# Core robotic dog dependencies
RPi.GPIO==0.7.1
gpiozero==1.6.2
picamera==1.13
opencv-python==4.8.1.78
numpy==1.24.3
websockets==11.0.3
asyncio-mqtt==0.13.0
flask==2.3.3
flask-socketio==5.3.6
requests==2.31.0
python-socketio==5.8.0
gpsd-py3==0.3.0
pyserial==3.5
adafruit-circuitpython-motor==3.4.8
adafruit-circuitpython-servokit==1.3.15
mediapipe==0.10.3